+++ src/components/gird/index.js
@@ -1,0 +1,5 @@
+import WGrid from "./src/grid";
+
+WGrid.install = (Vue) => {
+  Vue.component(WGrid.name, WGrid);
+};