+++ src/components/grid-item/index.js
@@ -1,0 +1,5 @@
+import WGridItem from "./src/grid-item";
+
+WGridItem.install = (Vue)=>{
+  Vue.component(WGridItem.name,WGridItem);
+};